[FEATURE] Allow appending "type" param to url
When adding a query param to a query-free SEF url dynamically via javascript it broke the page cache hash calculation in the TYPO3 bootstrap process as - in short - it does not recalculate the cHash but rather excludes the extension specific GET params from the page cache hash calculation. Therefore the cHash param should never get removed from the SEF url even if it is the only GET param.